extract plain old php classes to core
